VIX-1860 Add some logic to create backups of the core config files prior to saving over them. In the event of a save failure, the original file could be recovered and thus save the user from having to start over.

+++ Vixen.System/IO/Xml/XElementFileWriter.cs
@@ -1,24 +1,62 @@
 ﻿using System;
 using System.IO;
+using System.Linq;
 using System.Runtime.InteropServices;
+using System.Windows;
 using System.Xml.Linq;
 
 namespace Vixen.IO.Xml
 {
 	internal class XElementFileWriter : IFileWriter<XElement>
 	{
+		private static NLog.Logger Logging = NLog.LogManager.GetCurrentClassLogger();
 		public void WriteFile(string filePath, XElement content)
 		{
 		    while (IsFileLocked(filePath))
 		    {
+				Logging.Warn("Filepath {0} is locked! Sleeping for 250ms to wait for it to free up.", filePath);
 		        System.Threading.Thread.Sleep(250);
 		    }
-			content.Save(filePath);
+			try
+			{
+				content.Save(filePath);
+			}
+			catch (Exception e)
+			{
+				Logging.Error("An error occured trying to save the file. Attempting to protect any backups", e);
+
+				try
+				{
+					var backupFile = string.Format("{0}_{1}", filePath, "backup");
+					if (File.Exists(backupFile))
+					{
+						File.Copy(backupFile, string.Format("{0}.protected.{1}", filePath, DateTime.Now.ToFileTime()));
+					}
+				}
+				catch (Exception e2)
+				{
+					Logging.Error("Could not protect the backup file.", e2);
+				}
+
+				throw;
+			}
 		}
 
 		void IFileWriter.WriteFile(string filePath, object content)
 		{
 			if (!(content is XElement)) throw new InvalidOperationException("Content mst be an XElement.");
+			
+			var success = BackupFile(filePath);
+			if (!success)
+			{
+				var result = MessageBox.Show("Config backups failed prior to save! Possible file system issue!", "Warning!",
+					MessageBoxButton.OKCancel);
+				if (result == MessageBoxResult.Cancel)
+				{
+					return;
+				}
+			}
+
 			WriteFile(filePath, (XElement) content);
         }
         bool IsFileLocked(string path)
@@ -32,10 +70,37 @@ namespace Vixen.IO.Xml
             }
             catch (IOException e)
             {
+				Logging.Error("IO Exception checking file lock.", e);
                 return (Marshal.GetHRForException(e) & 0xFFFF) == 32;
             }
 
             return false;
         }
+
+		private bool BackupFile(string filePath)
+		{
+			bool success = false;
+			while (IsFileLocked(filePath))
+			{
+				Logging.Warn("Filepath {0} is locked! Sleeping for 250ms to wait for it to free up.", filePath);
+				System.Threading.Thread.Sleep(250);
+			}
+
+			try
+			{
+				if (File.Exists(filePath))
+				{
+					File.Copy(filePath, string.Format("{0}_{1}", filePath, "backup"), true);
+				}
+				success = true;
+			}
+			catch (Exception e)
+			{
+				Logging.Error("Error while backing up file!", e);
+			}
+
+			return success;
+
+		}
     }
 }
